beginning edits to rest node
