+++ default.node.info.yaml
@@ -1,0 +1,130 @@
+node_name: default.node
+node_id: 01K8C21HMA4JD5T4CS5W25VF3R
+node_description: null
+node_type: device
+module_name: liconic_rest_node
+module_version: 0.0.1
+capabilities:
+  get_info: true
+  get_state: true
+  get_status: true
+  send_action: true
+  get_action_result: true
+  get_action_history: true
+  action_files: true
+  send_admin_commands: true
+  set_config: true
+  get_resources: false
+  get_log: true
+  admin_commands:
+  - lock
+  - reset
+  - shutdown
+  - unlock
+node_url: null
+actions:
+  load_plate:
+    name: load_plate
+    description: Load a plate into the incubator
+    args:
+      stack:
+        name: stack
+        description: stack number (1-4)
+        argument_type: int
+        required: false
+        default: null
+      slot:
+        name: slot
+        description: slot number (1-22 for stacks 1 and 2, 1-10 for stacks 3 and 4)
+        argument_type: int
+        required: false
+        default: null
+    locations: {}
+    files: {}
+    results: {}
+    blocking: false
+    asynchronous: true
+config:
+  node_definition: default.node.yaml
+  node_info_path: null
+  update_node_files: true
+  status_update_interval: 2.0
+  state_update_interval: 2.0
+  node_url: http://127.0.0.1:2000/
+  uvicorn_kwargs: {}
+  port: 3333
+  host: localhost
+  resources_path: C:\Users\svcaibio\.madsci\liconic\liconic_resources.yaml
+config_schema:
+  description: Configuration for the Liconic REST node
+  properties:
+    node_definition:
+      anyOf:
+      - type: string
+      - format: path
+        type: string
+      - type: 'null'
+      default: default.node.yaml
+      description: Path to the node definition file to use. If set, the node will
+        load the definition from this file on startup. Otherwise, a default configuration
+        will be created.
+      title: Node Definition File
+    node_info_path:
+      anyOf:
+      - type: string
+      - format: path
+        type: string
+      - type: 'null'
+      default: null
+      description: Path to export the generated node info file. If not set, will use
+        the node name and the node_definition's path.
+      title: Node Info Path
+    update_node_files:
+      default: true
+      description: Whether to update the node definition and info files on startup.
+        If set to False, the node will not update the files even if they are out of
+        date.
+      title: Update Node Files
+      type: boolean
+    status_update_interval:
+      anyOf:
+      - type: number
+      - type: 'null'
+      default: 2.0
+      description: The interval in seconds at which the node should update its status.
+      title: Status Update Interval
+    state_update_interval:
+      anyOf:
+      - type: number
+      - type: 'null'
+      default: 2.0
+      description: The interval in seconds at which the node should update its state.
+      title: State Update Interval
+    node_url:
+      default: http://127.0.0.1:2000/
+      description: The URL used to communicate with the node. This is the base URL
+        for the REST API.
+      format: uri
+      minLength: 1
+      title: Node URL
+      type: string
+    uvicorn_kwargs:
+      additionalProperties: true
+      description: Configuration for the Uvicorn server that runs the REST API.
+      title: Uvicorn Configuration
+      type: object
+    port:
+      default: 3333
+      title: Port
+      type: integer
+    host:
+      default: localhost
+      title: Host
+      type: string
+    resources_path:
+      default: C:\Users\svcaibio\.madsci\liconic\liconic_resources.yaml
+      format: path
+      title: Resources Path
+      type: string
+  title: LiconicNodeConfig
+  type: object
