+++ default.node.yaml
@@ -1,0 +1,7 @@
+node_name: default.node
+node_id: 01K8C21HMA4JD5T4CS5W25VF3R
+node_description: null
+node_type: device
+module_name: liconic_rest_node
+module_version: 0.0.1
+capabilities: null
