+++ definitions/liconic_lisa.node.yaml
@@ -5,3 +5,6 @@ node_type: device
 module_name: liconic_module
 module_version: 0.4.0
 capabilities: null
+config_defaults:   # seemingly does nothing ....
+  port: 2003
+  host: hudson01.cels.anl.gov
